+++ CHANGELOG.md
@@ -1,13 +1,29 @@
 # MuxMate Changelog
 
-## 3.1.0 - 2025-03-10
+## 3.2.0
+
 ### Added
+
+- Basic GraphQL Types
+
+### Removed
+
+- Static MP4 generation
+
+## 3.1.0 - 2025-03-10
+
+### Added
+
 - Added support for rendering Twig in the `scriptSrcNonce` setting
 
 ## 3.0.1 - 2024-08-22
+
 ### Fixed
+
 - Fixed some minor Craft 5 compatibility issues
 
 ## 3.0.0 - 2024-08-22
+
 ### Added
+
 - Added support for Craft 5